backend/pyproject.toml
[tool.poetry]
name = "equipe-2"
version = "0.1.0"
description = ""
authors = ["João Pedro Gonçalves <45002908+jpedro-fsgs@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
sqlalchemy = "^2.0.32"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
pywhatkit = "^5.4"
crcmod = "^1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
